requirements.txt
# Core conversion stack
PyMuPDF>=1.23.8
python-docx>=0.8.11
pytesseract>=0.3.10
Pillow>=10.0.0

# Dev / testing (optional)
pytest>=7.4.0

# Packaging (optional; for later)
pyinstaller>=6.0.0
